+++ web/tomato/js/editor/components/elements/icon_elements/vm_element.js
@@ -5,6 +5,9 @@ var VMElement = IconElement.extend({
 		return this._super() && !this.busy;
 	},
 	iconUrl: function() {
+		if (this.data._custom_icon != undefined && this.data._custom_icon != null && this.data._custom_icon != "") {
+			return this.data._custom_icon;
+		}
 		return this.getTemplate() ? this.getTemplate().iconUrl() : this._super(); 
 	},
 	isRemovable: function() {
@@ -33,7 +36,7 @@ var VMElement = IconElement.extend({
 	},
 	configWindowSettings: function() {
 		var config = this._super();
-		config.order = ["name", "site", "profile", "template", "_endpoint"];
+		config.order = ["name", "site", "profile", "template", "_endpoint", "_custom_icon"];
 		config.ignore.push("info_last_sync");
 		
 		var profileInfo = {};
@@ -156,7 +159,13 @@ var VMElement = IconElement.extend({
 			choices: {true: "Seperates segments", false: "Connects segments"},
 			value: this.isEndpoint(),
 			inputConverter: Boolean.parse
-		}); 
+		});
+		config.special._custom_icon = new TextElement({
+			label: "Custom icon",
+			name: "_custom_icon",
+			value:this.data._custom_icon,
+			hint: "URL to 32x32 PNG image"
+		});
 		return config;
 	},
 	getConnectTarget: function(callback) {