+++ CHANGELOG.md
@@ -18,10 +18,12 @@ Entries should be sorted by component and status with important entries and entr
 - [backend_core, changed] Better timeout notifications and times
 - [config, changed] In dumpmanager config section, api_store_secret_key must now be set.
 - [config, changed] web module now requires a simple dump config
+- [config, changed] web/web-resources now requires the __custom-element-icons__ key 
 - [docker, changed] web docker image now supports a log directory
 - [backend_api, backend_debug, changed] Now allowing dump storing via API
 - [web, changed] web now sends error dumps to backend via API
 - [backend_users, fixed] Logins are now saved again for display in 30-day active users count
+- [web, changed] editor now allows to set custom icons for elements
 - [backend_api, fixed] Fixed __statistics__ API call
 - [backend_api, fixed] Fixed authorization for download actions on elements
 - [backend_api, changed] Removed configuration info calls